% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_names}
\alias{example_names}
\title{Example names}
\format{
vector of strings
}
\source{
George Lucas' brain \url{https://starwars.com}
}
\usage{
example_names
}
\description{
Example names from the Starwars universe.
}
\keyword{datasets}
