% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Make groups of people}
\usage{
make_groups(names)
}
\arguments{
\item{names}{a vector with names of people to match}
}
\value{
a matrix of coupled names
}
\description{
This function randomly matches a group of people. If the group is uneven,
then one person will be removed.
}
